Flatten `Computation` classes
