+++ torch_xla/csrc/runtime/compat_logging.h
@@ -1,0 +1,3 @@
+#include "torch_xla/csrc/runtime/tf_logging.h"
+#define C10_UTIL_LOGGING_IS_NOT_GOOGLE_GLOG_H_
+#include <c10/util/Logging.h>
